+++ dart/geobase/CHANGELOG.md
@@ -2,6 +2,9 @@
 
 NOTE: [geodata version 1.5.0](https://github.com/navibyte/geospatial/milestone/8) currently under development (1.5.0-dev.0).
 
+🧩 Features:
+* [Port conversions between (historical) ellipsoidal datums to Dart #256](https://github.com/navibyte/geospatial/issues/256)
+
 ## 1.4.0+1
 
 Small documentation link fixes.